Added problem statement as comment
There are 1000 doors in a high school with 1000 students. 
The problem begins with the first student opening all 1000 doors.
next the second student closes doors 2,4,6,8,10 and so on to door 
1000 the third student changes the state (opens doors closed, closes doors s open)
on doors 3,6,9,12,15 and so on the fourth student Changes the state of doors 4, 8, 12, 16 and so on.
This goes on until every student has had a turn. How many doors will be open at the end?